+++ src/main/java/br/edu/iff/ccc/librekan/librekan/service/QuadroService.java
@@ -1,5 +1,7 @@
 package br.edu.iff.ccc.librekan.librekan.service;
 
+import br.edu.iff.ccc.librekan.librekan.exceptions.RecursoNaoEncontradoException;
+import br.edu.iff.ccc.librekan.librekan.exceptions.RegraDeNegocioException;
 import br.edu.iff.ccc.librekan.librekan.model.Quadro;
 import br.edu.iff.ccc.librekan.librekan.repository.QuadroRepository;
 import org.springframework.stereotype.Service;
@@ -34,17 +36,19 @@ public class QuadroService {
     }
 
     public void excluir(Long id) {
+        if (quadroRepository.findById(id).isEmpty()) {
+            throw new RecursoNaoEncontradoException("Quadro não encontrado com o ID: " + id);
+        }
         quadroRepository.deleteById(id);
-
     }
     
-    @Transactional
+     @Transactional
     public Quadro atualizarNome(Long id, String novoNome) {
         if (quadroRepository.findByNomeIgnoreCaseAndIdNot(novoNome, id).isPresent()) {
-            throw new IllegalArgumentException("O nome do quadro '" + novoNome + "' já está em uso.");
+            throw new RegraDeNegocioException("O nome do quadro '" + novoNome + "' já está em uso.");
         }
         Quadro quadro = quadroRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Quadro não encontrado com o id: " + id));
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Quadro não encontrado com o id: " + id));
         
         quadro.setNome(novoNome);
         return quadroRepository.save(quadro);
